[FLINK-30790] Change Postgres tests to use new PostgresDatabase
