+++ flink-connector-jdbc/src/test/java/org/apache/flink/connector/jdbc/databases/postgres/PostgresDatabase.java
@@ -19,12 +19,14 @@ package org.apache.flink.connector.jdbc.databases.postgres;
 
 import org.apache.flink.connector.jdbc.databases.DatabaseMetadata;
 import org.apache.flink.connector.jdbc.databases.DatabaseTest;
-import org.apache.flink.connector.jdbc.dialect.postgres.PostgresExactlyOnceSinkE2eTest;
 import org.apache.flink.connector.jdbc.test.DockerImageVersions;
 
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+import static org.apache.flink.util.Preconditions.checkArgument;
 
 /** A Postgres database for testing. * */
 @Testcontainers
@@ -32,12 +34,52 @@ public interface PostgresDatabase extends DatabaseTest {
 
     @Container
     PostgreSQLContainer<?> CONTAINER =
-            new PostgresExactlyOnceSinkE2eTest.PostgresXaContainer(DockerImageVersions.POSTGRES)
+            new PostgresXaContainer(DockerImageVersions.POSTGRES)
                     .withMaxConnections(10)
                     .withMaxTransactions(50);
 
     @Override
     default DatabaseMetadata getMetadata() {
         return new PostgresMetadata(CONTAINER);
+    }
+
+    /** {@link PostgreSQLContainer} with XA enabled (by setting max_prepared_transactions). */
+    class PostgresXaContainer extends PostgreSQLContainer<PostgresXaContainer> {
+        private static final int SUPERUSER_RESERVED_CONNECTIONS = 1;
+        private int maxConnections = SUPERUSER_RESERVED_CONNECTIONS + 1;
+        private int maxTransactions = 1;
+
+        public PostgresXaContainer(String dockerImageName) {
+            super(DockerImageName.parse(dockerImageName));
+        }
+
+        public PostgresXaContainer withMaxConnections(int maxConnections) {
+            checkArgument(
+                    maxConnections > SUPERUSER_RESERVED_CONNECTIONS,
+                    "maxConnections should be greater than superuser_reserved_connections");
+            this.maxConnections = maxConnections;
+            return this.self();
+        }
+
+        public PostgresXaContainer withMaxTransactions(int maxTransactions) {
+            checkArgument(maxTransactions > 1, "maxTransactions should be greater 1");
+            this.maxTransactions = maxTransactions;
+            return this.self();
+        }
+
+        @Override
+        public void start() {
+            setCommand(
+                    "postgres",
+                    "-c",
+                    "superuser_reserved_connections=" + SUPERUSER_RESERVED_CONNECTIONS,
+                    "-c",
+                    "max_connections=" + maxConnections,
+                    "-c",
+                    "max_prepared_transactions=" + maxTransactions,
+                    "-c",
+                    "fsync=off");
+            super.start();
+        }
     }
 }
